refactor(scan): remove v0 ballot encoding
I initially kept v0 and v1 encodings so that we could scan in a backward-compatible way. However, in practice this has not been useful and has made things a little more confusing in development. Rather than maintain support for both, I'm removing v0 and only supporting v1. This is a BREAKING CHANGE, but since we're not actually publishing `@votingworks/ballot-encoder` to npm anymore this doesn't really matter.

This paves the way to break compatibility yet again by adding a partial election hash to the encoded ballot.
